dns-tunnel: base64 (not hex) pubkey docs; drop dead salt from MTU; fix 32-bit CIDR truncation (PR #48 review round 3)
Address Copilot's round-3 pass on #48.

Correctness:
- balancer.rs: CIDR host-count `(count as usize).min(max_hosts)` truncated
  `1u64 << 32` to 0 on 32-bit targets (e.g. Android armv7) → empty resolver
  pool for a large CIDR. Compute the min in u64 before the cast.
- mtu.rs: `wire_overhead`/`max_uplink_payload` still subtracted `SALT_LEN` for a
  "long-form SYN", but the forward-secret handshake removed the salt (handshake
  packets are cleartext FORM_SYN/FORM_SYNACK, sized separately) and every
  production caller already passed `long=false`. Drop the dead `long` param +
  salt term (and the obsolete test assertion); update the sole caller.

Docs / comments:
- Fix a base64/hex error I introduced last round: `server_pubkey` is **base64**
  (`crypto::decode_server_pub` + the real `DnsTunnelConfig` doc), not hex —
  corrected the design-doc config example + validation bullet and the live-test
  env-var comment (`DNS_TUNNEL_PUBKEY=<base64 Ed25519>`).
- frame.rs: `Kind::Syn`/`SynAck` doc said "session open (long form, salt,
  cookie)"; they are the sealed *stream*-open/ack (the session handshake is the
  cleartext FORM_SYN/FORM_SYNACK wire packets).
- transport/mod.rs: `spec_label` now says "N configured resolvers" — "0
  resolvers" was misleading given auto-included system resolvers / authoritative
  mode.

core 58 + server tests green; clippy/fmt clean; `--features dns-tunnel` builds.

+++ core/src/transport/dns_tunnel/balancer.rs
@@ -387,7 +387,7 @@ fn parse_spec(spec: &str, max_hosts: usize) -> Result<Vec<SocketAddr>, PoolError
         }
         let count = 1u64 << (32 - prefix);
         let base_u = u32::from(base) & (!0u32).checked_shl(32 - prefix).unwrap_or(0);
-        let n = (count as usize).min(max_hosts);
+        let n = count.min(max_hosts as u64) as usize;
         let mut out = Vec::with_capacity(n);
         for i in 0..n as u64 {
             let ip = Ipv4Addr::from(base_u.wrapping_add(i as u32));
